refactor: move CombatService to dedicated service package for better code hygiene
