+++ src/test/kotlin/at/aau/hexabrawl/websocketserver/TestServiceFactory.kt
@@ -1,6 +1,6 @@
 package at.aau.hexabrawl.websocketserver
 
-import at.aau.hexabrawl.websocketserver.model.CombatService
+import at.aau.hexabrawl.websocketserver.service.CombatService
 import at.aau.hexabrawl.websocketserver.model.GameService
 import at.aau.hexabrawl.websocketserver.model.ConnectivityService
